move SunCalc lib to androidApp for now
