+++ app/shared/build.gradle.kts
@@ -26,7 +26,6 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.4.0")
-                implementation("org.shredzone.commons:commons-suncalc:3.7")
             }
         }
         val commonTest by getting {
@@ -34,7 +33,11 @@ kotlin {
                 implementation(kotlin("test"))
             }
         }
-        val androidMain by getting
+        val androidMain by getting {
+            dependencies {
+                // todo suncalc
+            }
+        }
         val androidUnitTest by getting
         val iosX64Main by getting
         val iosArm64Main by getting
