--- app/shared/src/commonMain/kotlin/tt/co/jesses/moonlight/data/model/MoonIlluminationData.kt
@@ -1,10 +1,0 @@
-package tt.co.jesses.moonlight.data.model
-
-/**
- * Raw data representation of Moon illumination
- */
-data class MoonIlluminationData(
-    val fraction: Double,
-    val phase: Double,
-    val angle: Double,
-)
